+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.5.0]
-### Added pre_install and post_install hooks. Hooks are run at a per chart level
+### Added 
+- pre_install and post_install hooks. Hooks are run at a per chart level
 
 ## [0.3.0]
 ### Added